requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep Learning (optional, for advanced models)
# torch>=1.10.0
# tensorflow>=2.8.0

# NLP Processing
nltk>=3.6
textblob>=0.17.1

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
joblib>=1.1.0
pyyaml>=6.0
tqdm>=4.62.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=4.0.0
